chore(scripts): update license check to be more flexible @W-19066903 (#5422)
* chore(scripts): convert license script to ESM

* chore(scripts): change license script from .js to .mjs

* chore(scripts): make license file look in more places

@parse5/tools moved from root to template-compiler

and there's additional ESM vs CJS nonsense to deal with

* chore: add a comment
