+++ .git2gus/config.json
@@ -1,5 +1,5 @@
 {
     "productTag": "a1aB0000000g309IAA",
-    "defaultBuild": "256",
+    "defaultBuild": "258",
     "hideWorkItemUrl": true
 }
